fixed bugs ( detroy already detroyed buffer object, clear texture views when FrameBuffer is destroyed ).

+++ Doom3/Source/Core/Asset/TextureAsset.cpp
@@ -193,8 +193,11 @@ void dooms::asset::TextureAsset::AllocateTextureResourceObject()
 
 void dooms::asset::TextureAsset::DestroyTextureResourceObject()
 {
-	graphics::GraphicsAPI::DestroyTextureObject(mTextureResourceObject);
-	mTextureResourceObject.Reset();
+	if(mTextureResourceObject.IsValid())
+	{
+		graphics::GraphicsAPI::DestroyTextureObject(mTextureResourceObject);
+		mTextureResourceObject.Reset();
+	}
 }
 
 dooms::asset::eAssetType dooms::asset::TextureAsset::GetEAssetType() const
